refactor asset integration classes
