+++ care/utils/assetintegration/onvif.py
@@ -1,6 +1,4 @@
 import enum
-import json
-from urllib import request
 
 from care.utils.assetintegration.base import BaseAssetIntegration
 
@@ -14,33 +12,13 @@ class OnvifAsset(BaseAssetIntegration):
 
     def __init__(self, meta):
         try:
-            self.meta = json.loads(meta)
-            self.meta = meta
-            self.name = meta["camera_type"]
-            self.host = meta["local_ip_address"]
-            self.port = meta["camera_port"] or 80
-            self.username = meta["camera_access_key"].split(":")[0]
-            self.password = meta["access_credentials"].split(":")[1]
-            self.middleware_hostname = meta["middleware_hostname"]
+            super().__init__(meta)
+            self.name = self.meta["camera_type"]
+            self.port = self.meta["camera_port"] or 80
+            self.username = self.meta["camera_access_key"].split(":")[0]
+            self.password = self.meta["access_credentials"].split(":")[1]
         except KeyError:
             print("Error: Invalid Onvif Asset; Missing required fields")
-
-    def get_url(self, endpoint):
-        return "http://{}{}".format(self.middleware_hostname, endpoint)
-
-    def api_post(self, url, data=None):
-        req = request.post(url, json=data)
-        try:
-            return req.json()
-        except json.decoder.JSONDecodeError:
-            return {"error": "Invalid Response"}
-
-    def api_get(self, url, data=None):
-        req = request.get(url, data=data)
-        try:
-            return req.json()
-        except json.decoder.JSONDecodeError:
-            return {"error": "Invalid Response"}
 
     def handle_action(self, action):
         if action.type == self.OnvifActions.MOVE_ABSOLUTE.value:
